no-systray-mode: make app-exit and minimize work [fix]
If the system tray is not used (currently only possible if the underlying
OS/Window-Manager is not supported by Javas SystemTray implementation),
make jitsi exit when clicking on the X window button.
Optionally (setting in the 'General' tab) allow to minimize
the main window instead of exiting.
This can be changed in the 'General' tab in the GUI settings.